+++ test/libsolidity/smtCheckerTests/typecast/string_literal_to_fixed_bytes_explicit.sol
@@ -1,0 +1,4 @@
+pragma experimental SMTChecker;
+contract SMT {
+	bytes32 constant internal NULL_BYTES32 = bytes32('');
+}