+++ Changelog.md
@@ -5,6 +5,7 @@ Compiler Features:
 
 Bugfixes:
  * SMTChecker: Fix lack of reporting potential violations when using only the CHC engine.
+ * SMTChecker: Fix internal error on conversion from string literal to byte.
  * Code generator: Fix missing creation dependency tracking for abstract contracts.